+++ src/Exceptions/OtpMismatchException.php
@@ -2,7 +2,7 @@
 
 namespace aliirfaan\LaravelSimpleOtp\Exceptions;
 
-class NotMatchException extends \UnexpectedValueException
+class OtpMismatchException extends \UnexpectedValueException
 {
     //
 }