+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com) and this project adheres to [Semantic Versioning](https://semver.org).
 
+# 6.0.0 - xxxx-xx-xx
+
+REmoved otp_does_expire, now always expires
+removed should encode, now always encodes
+REvmoed fillable digit, now has a simulated OTP code, more 
+Rename model to lso_otps
+added device_id for otp per device
+Added otp_meta
+Added correlation_id
+otp_type: numeric, alpha numeric
+
 ## 5.2.2 - 2025-10-17
 
 ### Added